+++ forge/src/main/java/io/github/flemmli97/improvedmobs/forge/config/ConfigSpecs.java
@@ -80,6 +80,8 @@ public class ConfigSpecs {
         public final ForgeConfigSpec.DoubleValue playerEXScale;
         public final ForgeConfigSpec.EnumValue<Config.IntegrationType> useLevelZMod;
         public final ForgeConfigSpec.DoubleValue levelZScale;
+        public final ForgeConfigSpec.EnumValue<Config.IntegrationType> useRunecraftoryMod;
+        public final ForgeConfigSpec.DoubleValue runecraftoryScale;
         public final ForgeConfigSpec.BooleanValue varySizebyPehkui;
         public final ForgeConfigSpec.DoubleValue sizeMin;
         public final ForgeConfigSpec.DoubleValue sizeMax;
@@ -199,8 +201,10 @@ public class ConfigSpecs {
             this.useScalingHealthMod = builder.comment("Should the scaling health mods difficulty system be used instead of this ones. (Requires scaling health mod)").defineEnum("Use Scaling Health Mod", Config.CommonConfig.useScalingHealthMod);
             this.usePlayerEXMod = builder.comment("If true and playerEx is installed will use the level from playerEx as difficulty").defineEnum("Use Player EX Mod", Config.CommonConfig.usePlayerEXMod);
             this.playerEXScale = builder.comment("Scaling for playerEX integration").defineInRange("PlayerEX Scaling", Config.CommonConfig.playerEXScale, 0, Double.MAX_VALUE);
-            this.useLevelZMod = builder.comment("If true and LevelZ is installed will use the the total skill level from LevelZ as difficulty").defineEnum("Use LevelZ Mod", Config.CommonConfig.useLevelZMod);
+            this.useLevelZMod = builder.comment("If true and LevelZ is installed will use the total skill level from LevelZ as difficulty").defineEnum("Use LevelZ Mod", Config.CommonConfig.useLevelZMod);
             this.levelZScale = builder.comment("Scaling for LevelZ integration").defineInRange("LevelZ Scaling", Config.CommonConfig.levelZScale, 0, Double.MAX_VALUE);
+            this.useRunecraftoryMod = builder.comment("If true and Runecraftory is installed will use the level from Runecraftory as difficulty").defineEnum("Use Runecraftory Mod", Config.CommonConfig.useRunecraftoryMod);
+            this.runecraftoryScale = builder.comment("Scaling for Runecraftory integration").defineInRange("Runecraftory Scaling", Config.CommonConfig.runecraftoryScale, 0, Double.MAX_VALUE);
             this.varySizebyPehkui = builder.comment("Using pehkui to vary the size of mobs").define("Use pehkui Mod", Config.CommonConfig.varySizebyPehkui);
             this.sizeMax = builder.comment("The Max scale of mobs. Range [1.0,10], default 2.0").defineInRange("Max size Multiplier", Config.CommonConfig.sizeMax, 1.0, 10.0);
             this.sizeMin = builder.comment("The Minimum scale of mobs. Range (0,1.0), default 0.5").defineInRange("Minimum size Multiplier", Config.CommonConfig.sizeMin, 0, 1.0);